% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_teams.R
\name{compare_teams}
\alias{compare_teams}
\title{A table with some statistics comparing two teams in a game}
\usage{
compare_teams(df, game)
}
\arguments{
\item{df}{A data frame with data from the rocket league stats api}

\item{game}{A string specifying the guid of the desired match to analyze.}
}
\value{
A table with statistics comparing the teams partaking in a game
}
\description{
A table with some statistics comparing two teams in a game
}
\examples{
compare_teams(rocket_data, game = "C34377FE11F153C39C443797FB07660C")
}
