% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speed_boost_plots.R
\name{speed_boost_plots}
\alias{speed_boost_plots}
\title{Line plots showing a players boost and speed for a game}
\usage{
speed_boost_plots(df, name, game)
}
\arguments{
\item{df}{A data frame with data from the rocket league stats api}

\item{name}{A string specifying the gamer-tag of the desired player}

\item{game}{A string specifying the guid of the desired match to analyze.}
}
\value{
Two line plots. One depicting a player's speed over the course of a  specified game. The other depicts the players boost total
}
\description{
Line plots showing a players boost and speed for a game
}
\examples{
speed_boost_plots(TDMZ443games, name = "TDMZ443", "D96D303C11F15617E1F59A8137901B26")
}
