% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_for_player.R
\name{stats_for_player}
\alias{stats_for_player}
\title{Calculates some basic aggregate stats for a specified player}
\usage{
stats_for_player(df, name, game = NULL)
}
\arguments{
\item{df}{A data frame with data from the rocket league stats api}

\item{name}{A string specifying the gamer-tag of the desired player}

\item{game}{Optional. A string specifying the guid of the desired match to analyze. Otherwise aggregates over all matches.}
}
\value{
Some aggregate statistics for the specified player
}
\description{
Calculates some basic aggregate stats for a specified player
}
\examples{
stats_for_player(TDMZ443games, name = "TDMZ443")

stats_for_player(TDMZ443games, name = "TDMZ443", "D96D303C11F15617E1F59A8137901B26")
}
